% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{aggregate}
\alias{aggregate}
\title{Aggregate}
\usage{
aggregate(obj, factor, aggtype)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{factor}{A numeric vector of length 1 or 2 indicating the aggregation factor in the x and y dimensions.}

\item{aggtype}{A character string indicating the aggregation type. See Details.}
}
\value{
Void.
}
\description{
Aggregates a VeloxRaster object to a lower resolution.
}
\details{
\code{aggtype} must be one of the following: "sum", "mean", "min", "max", "median".
}

