% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{as.RasterLayer}
\alias{as.RasterLayer}
\title{Cast a VeloxRaster band as a RasterLayer object}
\usage{
as.RasterLayer(obj, band = 1)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{band}{Integer indicating the RasterLayer band to be transformed.}
}
\value{
A RasterLayer object.
}
\description{
\code{as.RasterLayer} creates a RasterLayer object from a VeloxRaster band.
}

