% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{as.RasterStack}
\alias{as.RasterStack}
\title{Cast a VeloxRaster as a RasterStack object}
\usage{
as.RasterStack(obj)
}
\arguments{
\item{obj}{A VeloxRaster object.}
}
\value{
A RasterStack object.
}
\description{
\code{as.RasterStack} creates a RasterStack object from a VeloxRaster.
}

