% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\name{bg_intersects}
\alias{bg_intersects}
\title{Test whether two BoostGeometries objects intersect}
\usage{
bg_intersects(obj1, obj2)
}
\arguments{
\item{obj1}{A BoostGeometries object.}

\item{obj2}{A BoostGeometries object;}
}
\value{
A list, with list element i an integer vector with the indices j for which intersects(x[i],y[j]) is TRUE.
}
\description{
Tests whether two BoostGeometries objects intersect (element-wise).
}
\examples{
pts = boost(sf::st_sfc(sf::st_point(c(.5,.5)),
sf::st_point(c(1.5, 1.5)),
sf::st_point(c(2.5, 2.5))))
pol = boost(sf::st_sfc(sf::st_polygon(
list(rbind(c(0,0), c(2,0), c(2,2), c(0,2), c(0,0))))))
bg_intersects(pol, pts)

}

