% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\name{boost}
\alias{boost}
\title{Cast a sfc object as a BoostGeometries object}
\usage{
boost(x, ...)
}
\arguments{
\item{x}{An sfc object.}
}
\value{
A BoostGeometries object.
}
\description{
\code{boost} creates a BoostGeometries object from a sfc object.
}
\examples{
## Make sfc_POINT
sfc <- st_sfc(st_point(c(0,1)))
## Cast to BoostPoints
boostPoints <- boost(sfc)

}

