% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{class}
\name{BoostGrid-class}
\alias{BoostGrid-class}
\title{An S4 virtual class for storing Boost grids in C++}
\description{
This is a virtual class for storing Rcpp pointers to C++ grid objects.
}

