% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{class}
\name{BoostMultiLines-class}
\alias{BoostMultiLines-class}
\title{An S4 class for storing Boost multiline collections in C++}
\description{
This is a class for storing Rcpp pointers to C++ MultiLineCollection objects.
}

