% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{class}
\name{BoostMultiPoints-class}
\alias{BoostMultiPoints-class}
\title{An S4 class for storing Boost multipoint collections in C++}
\description{
This is a class for storing Rcpp pointers to C++ MultiPointCollection objects.
}

