% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{class}
\name{BoostObject-class}
\alias{BoostObject-class}
\title{A S4 class for storing Boost objects in C++}
\description{
This is a virtual class for storing Rcpp pointers to C++ GeometryCollection and GridCollection objects.
}
\section{Slots}{

\describe{
\item{\code{geomcollection}}{Rcpp pointer.}

\item{\code{crs}}{An object of class \code{sf::crs}, storing the coordinate reference system info.}

\item{\code{precision}}{A numeric scalar.}
}}

