% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{class}
\name{BoostPointGrid-class}
\alias{BoostPointGrid-class}
\title{An S4 class for storing Boost point grids in C++}
\description{
This is a class for storing Rcpp pointers to C++ PointGrid objects.
}

