% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{crop}
\alias{crop}
\title{Crop a VeloxRaster object}
\usage{
crop(obj, y, copy = FALSE)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{y}{An object from which an \code{extent} object can be extracted. Usually a Spatial* or Raster* object.}

\item{copy}{Boolean. If true, \code{crop} returns a new VeloxRaster object containing the cropped raster.
Otherwise the input VeloxRaster is cropped.}
}
\value{
If \code{copy=TRUE}: A VeloxRaster object.
}
\description{
\code{crop}s a VeloxRaster object
}
\details{
Crops a VeloxRaster object to the extent of y. Note that currently the
extent of y must overlap with the extent of x, otherwise an error is thrown.
}

