% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{extract}
\alias{extract}
\title{Extract}
\usage{
extract(obj, sp, fun)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{sp}{A SpatialPolygons* object.}

\item{fun}{An R function. See Details.}
}
\value{
A numeric matrix. One row per element in \code{sp}, one column per band in \code{obj}.
}
\description{
Extracts the values of all cells whose centerpoint is in \code{SpatialPolygons*} object
\code{sp} and applies R function \code{fun}.
}
\details{
\code{fun} must be an R function accepting a numeric vector as its sole input.
}

