% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{getCoordinates}
\alias{getCoordinates}
\title{Get coordinates}
\usage{
getCoordinates(obj)
}
\arguments{
\item{obj}{A VeloxRaster object.}
}
\value{
A numeric matrix.
}
\description{
Returns a matrix containing the x-y coordinates of all cell center points of a VeloxRaster.
}

