% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{im2col}
\alias{im2col}
\title{im2col}
\usage{
im2col(obj, wrow, wcol, band, padval = 0)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{wrow}{Patch size in the y dimension.}

\item{wcol}{Patch size in the x dimension.}

\item{band}{The band to be flattened.}

\item{padval}{A padding value.}
}
\value{
A numeric matrix with \code{wrow*wcol} columns.
}
\description{
Creates a matrix of flattened image patches from a VeloxRaster band.
Order is left-to-right, top-to-bottom.
}

