% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{methods}
\name{length,BoostGeometries-method}
\alias{length,BoostGeometries-method}
\title{BoostGeometries Length}
\usage{
\S4method{length}{BoostGeometries}(x)
}
\arguments{
\item{x}{A BoostGeometries object.}
}
\value{
An integer scalar.
}
\description{
Returns the length (number of Geometries) of a BoostGeometries object.
}

