% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{medianFocal}
\alias{medianFocal}
\title{Median focal}
\usage{
medianFocal(obj, wrow, wcol, bands = 1)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{wrow}{y dimension of filter. Must be uneven integer.}

\item{wcol}{x dimension of filter. Must be uneven integer.}

\item{bands}{Numeric vector indicating bands where filter is applied.}
}
\value{
Void.
}
\description{
Applies a median filter of dimension \code{wcol x wrow} to a VeloxRaster.
}
\details{
Padding is currently not implemented.
}

