% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{methods}
\name{plot,BoostGeometries,ANY-method}
\alias{plot,BoostGeometries,ANY-method}
\title{Plot BoostGeometries}
\usage{
\S4method{plot}{BoostGeometries,ANY}(x, y, ...)
}
\arguments{
\item{x}{A BoostGeometries object.}

\item{y}{Not used.}

\item{...}{Passed to \code{sf::plot}.}
}
\value{
Void.
}
\description{
Plot a BoostGeometries object using the \code{sf} plotting function.
}

