% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{rasterizePolygons}
\alias{rasterizePolygons}
\title{Rasterize Polygons}
\usage{
rasterizePolygons(obj, spdf, field, band = 1, background = NULL)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{spdf}{A SpatialPolygonsDataFrame object.}

\item{field}{A character string corresponding to the name of a numeric column in \code{spdf}.}

\item{band}{The band of \code{obj} where the rasterized values are written.}

\item{background}{Optional. A numeric value assigned to all background cells.}
}
\value{
Void.
}
\description{
Rasterizes a SpatialPolygonsDataFrame, i.e., assigns the values in the \code{field} column of the
SPDF to the raster cells intersecting with the respective SpatialPolygon.
}
\details{
Note that rasterization is performed sequentially. Hence, cells being contained by multiple polygons
are assigned the value of the last polygon in the \code{spdf} object.
}

