% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostgeom.R
\docType{methods}
\name{[,BoostMultiPolygons,ANY,ANY-method}
\alias{[,BoostMultiLines,ANY,ANY-method}
\alias{[,BoostMultiPoints,ANY,ANY-method}
\alias{[,BoostMultiPolygons,ANY,ANY-method}
\title{Subset a BoostGeometries object}
\usage{
\S4method{[}{BoostMultiPolygons,ANY,ANY}(x, i)

\S4method{[}{BoostMultiLines,ANY,ANY}(x, i)

\S4method{[}{BoostMultiPoints,ANY,ANY}(x, i)
}
\arguments{
\item{x}{A BoostGeometries object.}

\item{i}{An integer vector index.}
}
\value{
A BoostGeometries object.
}
\description{
Extract a subset of geometries from a BoostGeometries object.
}

