% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{sumFocal}
\alias{sumFocal}
\title{Sum focal}
\usage{
sumFocal(obj, weights, bands = 1)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{weights}{A numeric matrix of weights. Both dimensions must be uneven.}

\item{bands}{Numeric vector indicating bands where filter is applied.}
}
\value{
Void.
}
\description{
Applies a focal sum with weights matrix \code{weights} to a VeloxRaster.
}
\details{
Padding is currently not implemented.
}

