% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{velox}
\alias{velox}
\title{Create a VeloxRaster object}
\usage{
velox(x, extent = NULL, res = NULL, crs = NULL)
}
\arguments{
\item{x}{A RasterLayer, RasterStack, matrix, list of matrices, or text string pointing to a GDAL-readable file.}

\item{extent}{An \code{extent} object or a numeric vector of length 4. Required if \code{x} is a matrix or list
of matrices, ignored otherwise.}

\item{res}{The x and y resolution of the raster as a numeric vector of length 2.}
}
\value{
A VeloxRaster object.
}
\description{
\code{velox} creates a VeloxRaster object.
}
\details{
Creates a VeloxRaster object. Note that VeloxRaster objects are Reference Class objects and thus mutable.
Thus, the usual R copy on modify semantics do not apply.
}

