% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\docType{class}
\name{VeloxRaster-class}
\alias{VeloxRaster}
\alias{VeloxRaster-class}
\title{A Reference Class for velox rasters}
\description{
A Reference Class for velox rasters
}
\section{Fields}{

\describe{
\item{\code{vRaster}}{External pointer to C++ object of class 'vRaster'}

\item{\code{dim}}{Raster dimensions}

\item{\code{extent}}{Raster extent}

\item{\code{res}}{Raster resolution}

\item{\code{nbands}}{Number of raster bands}

\item{\code{crs}}{Coordinate reference system (Proj4 string)}
}}

