% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_cast.R
\name{VeloxRaster_as.matrix}
\alias{VeloxRaster_as.matrix}
\title{Cast a VeloxRaster band as a matrix}
\arguments{
\item{band}{Integer indicating the VeloxRaster band to be transformed.}
}
\value{
A matrix.
}
\description{
\code{as.matrix} creates a matrix from a VeloxRaster band.
}

