% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_cast.R
\name{VeloxRaster_as.RasterLayer}
\alias{VeloxRaster_as.RasterLayer}
\title{Cast a VeloxRaster band as a RasterLayer object}
\arguments{
\item{band}{Integer indicating the VeloxRaster band to be transformed.}
}
\value{
A RasterLayer object.
}
\description{
\code{as.RasterLayer} creates a RasterLayer object from a VeloxRaster band.
}

