% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_cast.R
\name{VeloxRaster_as.RasterStack}
\alias{VeloxRaster_as.RasterStack}
\title{Cast a VeloxRaster as a RasterStack object}
\value{
A RasterStack object.
}
\description{
\code{as.RasterStack} creates a RasterStack object from a VeloxRaster.
}

