% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_crop.R
\name{VeloxRaster_crop}
\alias{VeloxRaster_crop}
\title{Crop a VeloxRaster object}
\arguments{
\item{y}{An object from which an \code{extent} object can be extracted. Usually a Spatial* or Raster* object.}
}
\value{
Void.
}
\description{
\code{crop}s a VeloxRaster object
}
\details{
Crops a VeloxRaster object to the extent of y. Note that currently the
extent of y must overlap with the extent of x, otherwise an error is thrown.
}

