% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox.R
\name{VeloxRaster_drop}
\alias{VeloxRaster_drop}
\title{Delete a raster band from a VeloxRaster}
\arguments{
\item{bands}{Numeric vector containing IDs of bands to be dropped.}
}
\value{
Void.
}
\description{
Delete a raster band from a VeloxRaster
}

