% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_extract.R
\name{VeloxRaster_extract}
\alias{VeloxRaster_extract}
\title{Extract Values Given Polygons}
\arguments{
\item{sp}{A SpatialPolygons* object.}

\item{fun}{An R function. See Details.}
}
\value{
If \code{fun} is passed: A numeric matrix with one row per element in \code{sp}, one column per band in the VeloxRaster.

Otherwise: A list of numeric matrices, with one list element per element in \code{sp}.
Each matrix consists of one column per band in the VeloxRaster, one row per raster cell intersecting with the polygon.
}
\description{
Extracts the values of all cells whose centerpoint is in \code{SpatialPolygons*} object
\code{sp} and optionally applies R function \code{fun}.
}
\details{
If passed, \code{fun} must be an R function accepting a numeric vector as its first (and only mandatory) argument, and returning a scalar.
If \code{fun} is \code{NULL}, \code{extract} returns a list of matrices, each matrix containing the raster values intersecting with the respective polygon.
}
\examples{
## Make VeloxRaster with two bands
set.seed(0)
mat1 <- matrix(rnorm(100), 10, 10)
mat2 <- matrix(rnorm(100), 10, 10)
vx <- velox(list(mat1, mat2), extent=c(0,1,0,1), res=c(0.1,0.1),
            crs="+proj=longlat +datum=WGS84 +no_defs")
## Make SpatialPolygons
library(sp)
library(rgeos)
coord <- cbind(0.5, 0.5)
spoint <- SpatialPoints(coords=coord)
spols <- gBuffer(spgeom=spoint, width=0.5)
## Extract
vx$extract(sp=spols, fun=mean)

}

