% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox.R
\name{VeloxRaster_get_data_type}
\alias{VeloxRaster_get_data_type}
\title{Get data type of a VeloxRaster}
\value{
A character string denoting a GDAL data type.
}
\description{
Get data type of a VeloxRaster
}
\details{
Note that this method returns the data type of the raster, not the storage mode.
Except in special cases, velox stores all raster data as double precision matrices.
}

