% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_coordinates.R
\name{VeloxRaster_getCoordinates}
\alias{VeloxRaster_getCoordinates}
\title{Get coordinates}
\value{
A numeric matrix.
}
\description{
Returns a matrix containing the x-y coordinates of all cell center points of a VeloxRaster.
}

