% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_focal.R
\name{VeloxRaster_sumFocal}
\alias{VeloxRaster_sumFocal}
\title{Sum focal}
\arguments{
\item{weights}{A numeric matrix of weights. Both dimensions must be uneven.}

\item{bands}{Numeric vector indicating bands where filter is applied.}
}
\value{
Void.
}
\description{
Applies a focal sum with weights matrix \code{weights} to a VeloxRaster.
}
\details{
Padding is currently not implemented.
}

