% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox.R
\name{VeloxRaster_write}
\alias{VeloxRaster_write}
\title{Write a VeloxRaster to disk as a GeoTiff file}
\arguments{
\item{path}{Output filename as character string.}

\item{overwrite}{Boolean indicating whether target file should be overwritten.}
}
\value{
Void.
}
\description{
Write a VeloxRaster to disk as a GeoTiff file
}

