% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{write}
\alias{write}
\title{Write VeloxRaster to a file}
\usage{
write(obj, path, overwrite = FALSE)
}
\arguments{
\item{obj}{The VeloxRaster to write.}

\item{path}{Output filename as character string.}

\item{overwrite}{Boolean indicating whether target file should be overwritten.}
}
\value{
Void.
}
\description{
\code{write} writes a VeloxRaster object as a GeoTIFF file.
}
\details{
Use the '.tif' file ending.
}

