% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_1.0.R
\name{col2im}
\alias{col2im}
\title{col2im}
\usage{
col2im(obj, mat, wrow, wcol, band)
}
\arguments{
\item{obj}{A VeloxRaster object.}

\item{mat}{The matrix of flattened image patches.}

\item{wrow}{Patch size in the y dimension.}

\item{wcol}{Patch size in the x dimension.}

\item{band}{The band to be assigned.}
}
\value{
Void.
}
\description{
Assigns values to a VeloxRaster band from a matrix of flattened image patches.
}

