% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velox_coordinates.R
\name{VeloxRaster_getCoordinates}
\alias{VeloxRaster_getCoordinates}
\title{Get coordinates}
\value{
A numeric matrix.
}
\description{
Returns a matrix containing the x-y coordinates of all cell center points of a VeloxRaster.
}
\examples{
## Make VeloxRaster
mat <- matrix(1:100, 10, 10)
vx <- velox(mat, extent=c(0,1,0,1), res=c(0.1,0.1), crs="+proj=longlat +datum=WGS84 +no_defs")
## Get coordinate matrix
cmat <- vx$getCoordinates()

}

